// generated/api-client/models/my-blog-response-dto.ts
export interface MyBlogResponseDto {
    /**
     * 
     * @type {string}
     * @memberof MyBlogResponseDto
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof MyBlogResponseDto
     */
    'title': string;
    /**
     * 
     * @type {string}
     * @memberof MyBlogResponseDto
     */
    'slug': string;
    /**
     * 
     * @type {string}
     * @memberof MyBlogResponseDto
     */
    'subtitle': string | null;
    /**
     * 
     * @type {string}
     * @memberof MyBlogResponseDto
     */
    'excerpt': string | null;
    /**
     * 
     * @type {number}
     * @memberof MyBlogResponseDto
     */
    'readingTime': number | null;
    /**
     * 
     * @type {number}
     * @memberof MyBlogResponseDto
     */
    'wordCount': number | null;
    /**
     * 
     * @type {string}
     * @memberof MyBlogResponseDto
     */
    'status': string;
    /**
     * 
     * @type {string}
     * @memberof MyBlogResponseDto
     */
    'publishedAt': string | null;
    /**
     * 
     * @type {number}
     * @memberof MyBlogResponseDto
     */
    'viewCount': number;
    /**
     * 
     * @type {number}
     * @memberof MyBlogResponseDto
     */
    'likeCount': number;
    /**
     * 
     * @type {number}
     * @memberof MyBlogResponseDto
     */
    'commentCount': number;
    /**
     * 
     * @type {number}
     * @memberof MyBlogResponseDto
     */
    'bookmarkCount': number;
    /**
     * 
     * @type {number}
     * @memberof MyBlogResponseDto
     */
    'shareCount': number;
    /**
     * 
     * @type {boolean}
     * @memberof MyBlogResponseDto
     */
    'isPinned': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof MyBlogResponseDto
     */
    'isFeatured': boolean;
    /**
     * 
     * @type {string}
     * @memberof MyBlogResponseDto
     */
    'moderationStatus': string;
    /**
     * 
     * @type {boolean}
     * @memberof MyBlogResponseDto
     */
    'allowComments': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof MyBlogResponseDto
     */
    'allowBookmarks': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof MyBlogResponseDto
     */
    'allowLikes': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof MyBlogResponseDto
     */
    'allowShares': boolean;
    /**
     * 
     * @type {Array<MyBlogResponseDtoCoverImagesInner>}
     * @memberof MyBlogResponseDto
     */
    'coverImages': Array<MyBlogResponseDtoCoverImagesInner>;
    /**
     * 
     * @type {Array<MyBlogResponseDtoTagsInner>}
     * @memberof MyBlogResponseDto
     */
    'tags': Array<MyBlogResponseDtoTagsInner>;
    /**
     * 
     * @type {Array<MyBlogResponseDtoCategoriesInner>}
     * @memberof MyBlogResponseDto
     */
    'categories': Array<MyBlogResponseDtoCategoriesInner>;
    /**
     * 
     * @type {MyBlogResponseDtoCount}
     * @memberof MyBlogResponseDto
     */
    '_count': MyBlogResponseDtoCount;
    /**
     * 
     * @type {string}
     * @memberof MyBlogResponseDto
     */
    'createdAt': string;
    /**
     * 
     * @type {string}
     * @memberof MyBlogResponseDto
     */
    'updatedAt': string;
    /**
     * 
     * @type {string}
     * @memberof MyBlogResponseDto
     */
    'lastEditedAt': string;
}